Rename and move the object inspector component into a components folder
