+++ packages/devtools-reps/src/object-inspector/index.js
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
 
-const ObjectInspector = require("./component");
+const ObjectInspector = require("./components/ObjectInspector");
 const utils = require("./utils");
 const reducer = require("./reducer");
